집목록 테스트중

+++ korea_tour/src/main/java/com/bitcamp/korea_tour/model/homestay/HomeStayListDto.java
@@ -21,6 +21,6 @@ public class HomeStayListDto {
 	private List<HomeStayPhotoDto> homeStayPhotos;
 	private List<HomeStayMarkDto> homeStayMarkNums;
 	private int countOfReview;
-	private int avgOfStar;
+	private Double avgOfStar;
 	
 }
